+++ vercel.json
@@ -1,13 +1,4 @@
 {
   "version": 2,
-  "builds": [
-    {
-      "src": "index.js",
-      "use": "@vercel/node",
-      "files": ["userDB.json"]
-    }
-  ],
-  "routes": [
-    { "src": "/.*", "dest": "index.js" }
-  ]
+  "rewrites": [{ "source": "/", "destination": "/api" }]
 }
